requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
twilio
openai
langchain
langchain-community
eng-to-ipa
mysql-connector-python
azure-cognitiveservices-speech
azure-storage-blob